add spades cpp
